+++ packages/core/src/index.ts
@@ -6,6 +6,7 @@ export * from "./normalization.js";
 export * from "./model-assessment.js";
 export * from "./presentation.js";
 export * from "./roles.js";
+export * from "./route-policy.js";
 export * from "./routing.js";
 export * from "./service.js";
 export * from "./state-validation.js";